requirements.txt
Flask
flask-socketio
google-generativeai
python-dotenv
requests
